requirements.txt
azure-identity==1.16.0
azure-mgmt-resource==23.0.1
azure-mgmt-compute==30.6.0
azure-mgmt-storage==21.1.0
azure-mgmt-network==25.2.0
azure-monitor-query==1.3.0
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
apscheduler==3.10.4
